#15 correct hash tutorial code
