--- hash_proof/hash.py
@@ -1,14 +1,0 @@
-from hashlib import sha256
-
-
-def str_to_512bits(value: str) -> bytes:
-    bin_str: int = int(''.join(format(i, '08b') for i in value.encode("utf-8")), base=2)
-    padded_bytes: bytes = bin_str.to_bytes(64, "big")
-    return padded_bytes
-
-
-preimage: bytes = str_to_512bits("super secret value")
-hash: bytes = sha256(preimage).digest()
-
-print(f"preimage: 0x{preimage.hex()}")
-print(f"hash: 0x{hashed_msg.hex()}")
